fix: move AppsPermissions to apps-engine/definition
