+++ packages/apps-engine/src/definition/metadata/index.ts
@@ -4,6 +4,7 @@ import type { IAppInfo } from './IAppInfo';
 import { RocketChatAssociationModel, RocketChatAssociationRecord } from './RocketChatAssociations';
 
 export * from './AppInterface';
+export * from './AppPermissions';
 
 export type { IAppAuthorInfo, IAppInfo };
 export { AppMethod, RocketChatAssociationModel, RocketChatAssociationRecord };
